+++ glslang/MachineIndependent/SymbolTable.h
@@ -630,7 +630,7 @@ public:
     bool atGlobalLevel()  { return isGlobalLevel(currentLevel()); }
 
     int getGlobalLevel() { return globalLevel; }
-    int getCurrentLevelCount() { return table.size(); }
+    int getCurrentLevelCount() { return (int)(table.size()); }
 
     void setNoBuiltInRedeclarations() { noBuiltInRedeclarations = true; }
     void setSeparateNameSpaces() { separateNameSpaces = true; }
